requirements.txt
OpenAI
Flask
python-dotenv
chromadb
#chromadb.utils
sentence-transformers
flask-cors
APScheduler
PyPDF2 
pathlib
#pdfminer.six
#transformers
#torch
